TCAM/requirements.txt
lxml
matplotlib
numpy
tqdm
pycocotools
Pillow
torch==1.13.1
torchvision==0.14.1
